Add RoleService
Added the role service, updates roles on points. Problem with where we're updating though...
If the DB call only makes it half way, the user will never receive a role update. We need to have either
a) The role updates on a timer (which is server taxing)
or
b) Find a nice hook or instance to put into play on the main events and have a watcher.
I.e., every 10k messages sent in the server or something.